#34: Refactoring, adding TestDeleteRawData
